Refactor: 추천 검색어 리스트 api 호출 로직 customHook으로 분리
